kkjj-30 <refactor> 미완성 클래스 임시 비활성화 및 일부 삭제

+++ src/main/java/renewal/awesome_travel/event/dto/EventProductDto.java
@@ -21,7 +21,6 @@ public class EventProductDto {
 
     private Product product;
 
-    private Integer price;
-
+    private Long price;
 
 }
